tools: Introduce xmlgen.py
Add helper for generating cpuid XML bitfields.

Develop a proper data model so that the same code can also be used for
the XML parser ("<--" is inheritance, "<==" is composition):

  BitRange  <== CpuidBitRange, CpuidOSFeature
  CpuidNode <-- Cpuid{Leaf, Subleaf, Register, BitRange, OSFeature, ...}
  CpuidRegister: Insertion-ordered list of non-overlapping CpuidBitRange(s)

Let the Cpuid* objects link to each other in a tree form that mirrors
the XML schema:

  objtree = \
      CpuidLeaf('0x80000010', [
          CpuidSubleaf(0, [
              CpuidRegister('ebx', [
                  CpuidBitRange(BitRange(0, 1), desc="abc"),
                  CpuidBitRange(BitRange(3, 5), reserved=True),
                  CpuidBitRange(BitRange(7, 9), desc="xyz"),
              ])
          ]),
          CpuidSubleaf(1, [
              CpuidRegister('eax', [
                  CpuidBitRange(BitRange(3, 5), [
                      CpuidLinuxFeature(proc=False)
                  ])
              ])
          ])
      ])

This way, both the XML parser and the XML generator can operate on the
same data model using a basic visitor design pattern.

Signed-off-by: Ahmed S. Darwish <darwi@linutronix.de>