db: Introduce kcpuid.xslt
Transform the existing CPUID leaves XML database to a kcpuid-compatible
CSV file.

For the XSLT, uses a push-based instead of a pull-based approach;
i.e. "xsl:apply-templates" and "xsl:template" instead of "xsl:for-each"
with high levels of nesting. This makes the XSLT script more modular and
easier to maintain or modify in the future.

Note, kcpuid resides in the linux kernel source at tools/arch/x86/kcpuid.
The generated CSV file will be submitted to LKML in a later step.

Signed-off-by: Ahmed S. Darwish <darwi@linutronix.de>