LICENSES: Add the GPL-2.0-only and CC0-1.0 licenses
For REUSE tooling compliance, add the full text of the "GPL-2.0-only"
and "CC0-1.0" licenses to the source tree.  The former is used for
actual project components while the latter is for files that are not
particularly copyrightable, e.g. README and .gitignore, as REUSE
mandates that even such files must be marked with SPDX license tags.

The licenses' text was copied verbatim from SPDX's "license-list-data"
git repository.

With this commit, plus the SPDX license tags formerly added to all
project files, the x86-cpuid.org project is now fully compliant with
version 3.0 of the REUSE Specification.

(copy from external source:
 git: https://github.com/spdx/license-list-data/
 archive: 2dc9a03a1b8df042c683924bbd6b703ade448a0d [v3.22]
 copy: text/GPL-2.0-only.txt --> LICENSES/
 md5sum: 3d26203303a722dedc6bf909d95ba815
 copy: text/CC0-1.0.txt --> LICENSES/
 md5sum: 65d3616852dbf7b1a6d4b53b00626032)

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Ahmed S. Darwish <darwi@linutronix.de>
Link: https://reuse.software/
Link: https://github.com/fsfe/reuse-tool