pyproject.toml: dependencies: Add pygit2
The cpuidgen tool and its XSLT processors will add the project's git SHA
to all generated files.  Prepare for this by adding the pygit2 PIP as a
project dependency.

The first chosen candiate was the "python-git-info" PIP, mainly due to
its python-native git directory parsing capabilities relative to pygit2
which depends on libgit2.  Unfortunately the former did not have any
support for detecting whether a reporsitory is clean or dirty, which is a
required feature for the generated files.

Thus, depend on pygit2 until submitting patches to python-git-info
upstream.

Signed-off-by: Ahmed S. Darwish <darwi@linutronix.de>