kcpuid.xslt: Insert project version string
Inject a header at the generated cpuid.csv file showing the generator
project and its version tag.  The actual info must be passed by the
program invoking the XSLT's initial-template.

Modify README's "cpuidgen --kcpuid" example to include the newly
generated header.  The SPDX license tag was already generated by the
tool but not shown in the README's example, so add it as well.

Signed-off-by: Ahmed S. Darwish <darwi@linutronix.de>