kheader.xslt: Insert project version string
Similar to kcpuid.xslt, inject a header at the generated file showing
the generator project and its version string.  The actual info must be
passed by the program invoking the XSLT's initial-template.

Modify README's "cpuidgen --kheader" example to include the newly
generated header.  The SPDX license tag was already generated by the
tool but not shown in the README's example, so add it as well.

Signed-off-by: Ahmed S. Darwish <darwi@linutronix.de>