kheader-common.xslt: refactoring: Remove global variable
The $leafBiggestNameLen global variable was created for the C99 bitfield
variable names padding in the C structure.

It's calculated by checking all the bitfield variable names length,
finding the biggest length, and using that as the padding for the entire
leaf's header file (including subleaves).  This way, whatever comes
after the bitfield name is properly aligned across the whole header.

For example, cpuid leaf 0x12:

    struct leaf12_sl1 {
             // eax
             u32 secs_attr_init          :  1, //...
                 ...
                 secs_attr_einittoken_key:  1, //...
                 ...;
    };

    struct leaf12_sl2 {
        // eax
        u32     subleaf_type             :  4, //...
                                         :  8, // Reserved
                ...;
    };
                                         ^
                                         ^ Calculated padding

A new transformer will be added in further commits where the bitfield
descriptions for *all* cpuid leaves will be generated in one header
file.  Thus the padding value needs to be differentiated depending on
the user of kheader-common.xslt's templates and functions.

Remove the $leafBiggestNameLen global variable, and expect callers to
pass it instead.  Internally at kheader-common.xslt, trickle it down
across all templates and functions.

Signed-off-by: Ahmed S. Darwish <darwi@linutronix.de>