db: Use 'last=' for leaves with dynamic subleaf range
For leaves with a dynamic subleaf range, the 'array=' attribute is used
to denote the size of the subleaf storage.

Per latest LKML discussions, that 'array=' notation is confusing.  It
would be better to mark the dynamic subleaf range's 'first' and 'last'
possible subleaf instead.

Change the leaves XML descriptions to use 'first' and 'last'.  That is,
instead of:

    leaf_04.xml:    <subleaf id="0" array="32">
    leaf_0d.xml:    <subleaf id="2" array="62">
    ...

do:

    leaf_04.xml:    <subleaf id="0" last="31">
    leaf_0d.xml:    <subleaf id="2" last="63">
    ...

for all dynamic subleaf ranges.

Adjust the XSD schema and XSLT code accordingly.

Suggested-by: Sean Christopherson <seanjc@google.com>
Signed-off-by: Ahmed S. Darwish <darwi@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>
Link: https://lore.kernel.org/x86-cpuid/aKR3zjxsM7xQZ_g8@lx-t490